+++ actix-identity/tests/integration/test_app.rs
@@ -4,6 +4,7 @@ use actix_identity::{config::IdentityMiddlewareBuilder, Identity, IdentityMiddle
 use actix_session::{Session, SessionStatus};
 use actix_web::{web, App, HttpMessage, HttpRequest, HttpResponse, HttpServer};
 use serde::{Deserialize, Serialize};
+use serde_json::Value;
 
 use crate::fixtures::session_middleware;
 
@@ -136,7 +137,7 @@ async fn increment(session: Session, user: Option<Identity>) -> HttpResponse {
         .get::<i32>("counter")
         .unwrap_or(Some(0))
         .map_or(1, |inner| inner + 1);
-    session.insert("counter", counter).unwrap();
+    session.insert("counter", Value::from(counter));
 
     HttpResponse::Ok().json(&EndpointResponse {
         user_id,