+++ actix-session/src/storage/redis_actor.rs
@@ -276,9 +276,8 @@ impl SessionStore for RedisActorSessionStore {
 
 #[cfg(test)]
 mod tests {
-    use std::collections::HashMap;
-
     use actix_web::cookie::time::Duration;
+    use serde_json::Map;
 
     use super::*;
     use crate::test_helpers::acceptance_test_suite;
@@ -305,7 +304,7 @@ mod tests {
         let session_key = generate_session_key();
         let initial_session_key = session_key.as_ref().to_owned();
         let updated_session_key = store
-            .update(session_key, HashMap::new(), &Duration::seconds(1))
+            .update(session_key, Map::new(), &Duration::seconds(1))
             .await
             .unwrap();
         assert_ne!(initial_session_key, updated_session_key.as_ref());