+++ actix-session/tests/opaque_errors.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashMap, convert::TryInto};
+use std::convert::TryInto;
 
 use actix_session::{
     storage::{LoadError, SaveError, SessionKey, SessionStore, UpdateError},
@@ -12,6 +12,7 @@ use actix_web::{
     test, web, App, Responder,
 };
 use anyhow::Error;
+use serde_json::{Map, Value};
 
 #[actix_web::test]
 async fn errors_are_opaque() {
@@ -49,7 +50,7 @@ impl SessionStore for MockStore {
     async fn load(
         &self,
         _session_key: &SessionKey,
-    ) -> Result<Option<HashMap<String, String>>, LoadError> {
+    ) -> Result<Option<Map<String, Value>>, LoadError> {
         Err(LoadError::Other(anyhow::anyhow!(
             "My error full of implementation details"
         )))
@@ -57,7 +58,7 @@ impl SessionStore for MockStore {
 
     async fn save(
         &self,
-        _session_state: HashMap<String, String>,
+        _session_state: Map<String, Value>,
         _ttl: &Duration,
     ) -> Result<SessionKey, SaveError> {
         Ok("random_value".to_string().try_into().unwrap())
@@ -66,7 +67,7 @@ impl SessionStore for MockStore {
     async fn update(
         &self,
         _session_key: SessionKey,
-        _session_state: HashMap<String, String>,
+        _session_state: Map<String, Value>,
         _ttl: &Duration,
     ) -> Result<SessionKey, UpdateError> {
         #![allow(clippy::diverging_sub_expression)]
@@ -85,7 +86,7 @@ impl SessionStore for MockStore {
 }
 
 async fn create_session(session: Session) -> impl Responder {
-    session.insert("user_id", "id").unwrap();
+    session.insert("user_id", Value::from("id"));
     "Created"
 }