+++ actix-identity/CHANGES.md
@@ -2,9 +2,7 @@
 
 ## Unreleased
 
-## 0.7.0
-
-- Updated the usage of HashMap with serde Map to be compatible with `actix-session@0.9.0`.
+- Updated the usage of `HashMap` with `serde::Map` to be compatible with `actix-session` `0.9`.
 
 ## 0.6.0